Moved frontend files to subdirectory
