+++ frontend/package.json
@@ -8,9 +8,11 @@
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.2.2",
     "cdbreact": "^1.5.3",
+    "chart.js": "^3.9.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
     "react-calendar": "^3.9.0",
+    "react-chartjs-2": "^4.3.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
