+++ frontend/src/components/Historia.js
@@ -1,0 +1,10 @@
+
+const Historia = () => {
+    return(
+        <h2>Mahti hieno kaavio tähän</h2>
+    )
+
+}
+
+export default Historia;
+
