+++ frontend/src/components/HistoriaChart.js
@@ -1,0 +1,2 @@
+
+// Löydä chartit mitkä ei oo tehty typescriptalla
