+++ frontend/src/components/Kalenteri3.js
@@ -32,7 +32,7 @@ const [showTime, setShowTime] = useState(false)
    <Time showTime={showTime} date={date}/>
 
 
-						<Link to="/sivu2"><h1>Luo treeni</h1></Link>
+						<Link to={{pathname:"/sivu2", state:date}}><h1>Luo treeni</h1></Link>
 
 
  </div>
