+++ frontend/src/components/WorkoutFinished.js
@@ -1,0 +1,12 @@
+
+const WorkoutFinished = () => {
+
+    return(
+        <div>
+            <h1>Your workout is now finished and saved</h1>
+            <h2>Sending nukes to localstorage... You can start a new workout now.</h2>
+        </div>
+    )
+}
+
+export default WorkoutFinished;
