+++ frontend/src/components/sivu2.js
@@ -1,0 +1,17 @@
+import {Link} from 'react-router-dom';
+
+
+const Sivu2 = () => {
+return(
+    
+    <div>
+        <br/>
+        <h2>Tämä on sivu kaksi missä luodaan reeni</h2>
+
+		    <Link to="/Doingworkout"><h1>Tee treeni</h1></Link>
+
+    </div>
+)
+}
+
+export default Sivu2;
