--- src/components/Historia.js
@@ -1,10 +1,0 @@
-
-const Historia = () => {
-    return(
-        <h2>Mahti hieno kaavio tähän</h2>
-    )
-
-}
-
-export default Historia;
-
