--- src/components/OlioMockUp.js
@@ -1,9 +1,0 @@
-// Dataa varten olio pyörittelyä
-
-const [liike,setLiike] = useState({
-    type:"",
-    date:0,
-    weight:0,
-    sarjat:0
-})
-
